rename model_wrapper to model_openvino
